+++ app/controllers/api/v2/account_lists/mail_chimp_account_controller.rb
@@ -1,4 +1,6 @@
-class Api::V2::AccountLists::MailChimpAccountsController < Api::V2Controller
+class Api::V2::AccountLists::MailChimpAccountController < Api::V2Controller
+  resource_type :mail_chimp_account
+
   def show
     load_mailchimp_account
     authorize_mailchimp_account