+++ config/routes.rb
@@ -26,7 +26,7 @@ Rails.application.routes.draw do
               end
             end
 
-            resource :mail_chimp_account, only: [:show, :create, :destroy] do
+            resource :mail_chimp_account, only: [:show, :create, :destroy], controller: :mail_chimp_account do
               get :sync, on: :member
             end