+++ spec/controllers/api/v2/account_lists/mail_chimp_account_controller_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-describe Api::V2::AccountLists::MailChimpAccountsController, type: :controller do
+describe Api::V2::AccountLists::MailChimpAccountController, type: :controller do
   let(:factory_type) { :mail_chimp_account }
   let!(:user)        { create(:user_with_account) }