+++ CHANGELOG.md
@@ -6,6 +6,11 @@ This changelog covers what's changed in the MPDX APIs.
 - Added a serializer for Organization
 - Added a contact_id filter to Donations endpoint
 
+## 15 February 2017
+- Changed mail_chimp_accounts controller to mail_chimp_account
+- Added documentation for create action of account_lists#mail_chimp_account endpoint
+- Attempted to fix the seeder_application_spec intermittent error
+
 ## 13 February 2017
 - Nested Attributes with Included data can now be sent in POST/PATCH
 - Activity Comments are now officially Comments